refactor(card): rename poster to portrait
